fix: 스키마 구조와 불일치한 코드 정리

+++ src/main/java/com/zeropick/backend/product/entity/Product.java
@@ -15,7 +15,7 @@ public class Product {
     private Long id;
 
     private String name;
-    private Integer grade;
+    private Short grade;
     private Boolean warningAdditive;
     private Integer calories;
     private BigDecimal sugar;
